fix(Input.Number): Allow for floats
Input.Number allowed entering floating point numbers, but didn't allow to save them (because they would get converted to integers)

+++ specifyweb/frontend/js_src/lib/components/Atoms/Form.tsx
@@ -218,7 +218,7 @@ export const Input = {
       ...withHandleBlur(props.onBlur),
       onChange(event): void {
         onValueChange?.(
-          Number.parseInt((event.target as HTMLInputElement).value)
+          Number.parseFloat((event.target as HTMLInputElement).value)
         );
         props.onChange?.(event);
       },
